Fnc: Outil de purge de plages opératoires vides
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@23852 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPbloc/purge_empty_plagesop.php
@@ -33,9 +33,11 @@ if ($purge) {
   foreach ($plages as $_plage) {
 
     // Suppression des affectationde personnel
-    foreach ($_plage->loadAffectationsPersonnel() as $_affectations) {
-      foreach ($_affectations as $_affectation) {
-        $_affectation->delete();
+    if ($affectations = $_plage->loadAffectationsPersonnel()) {
+      foreach ($affectations as $_affectations) {
+        foreach ($_affectations as $_affectation) {
+          $_affectation->delete();
+        }
       }
     }
